refactor(routes): extract /bugs into its own module (god-file decomposition)
Bug routes moved verbatim from mycelium.js into server/routes/bugs.js (handler bodies
byte-identical; POST /bugs's agentWriteLimiter middleware preserved in position), wired
via registerBugRoutes. Route contract unchanged (284 routes incl. middleware counts),
vitest 265/265. Lucy (GLM-5.2), byte-reviewed by m5Max.
